+++ cogment/__init__.py
@@ -26,6 +26,7 @@ framework python SDK by AI Redefined Inc.
 from cogment.version import __version__
 
 from cogment.context import Context, Endpoint, ServedEndpoint
+from cogment.session import EventType
 
 # Necessary because of cogment CLI "cog_settings.py" generated code
 from cogment.actor import ActorClass, ActorClassList